feat(adieu-razzle): reusing existing code
